feat: file templates for scheduled bot
